pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["app/babel_tower"]

[tool.hatch.build.targets.wheel]
packages = ["app/babel_tower"]

[tool.hatch.build.targets.wheel.force-include]
"devops.py" = "devops.py"
"prompts" = "babel_tower/prompts"

[tool.hatch.build.targets.wheel.sources]
"app" = ""

[project]
name = "babel-tower"
version = "0.1.0"
description = "Voice input pipeline for Claude Code — STT with LLM postprocessing"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastmcp>=3.0.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.7.0",
    "silero-vad>=5.1",
    "onnxruntime>=1.19.0",
    "sounddevice>=0.5.0",
    "soundfile>=0.13.0",
    "numpy>=2.0.0",
    "typer>=0.15.0",
    "uvicorn>=0.34.0",
    "starlette>=0.45.0",
    "loguru>=0.7.0",
    "python-telegram-bot[socks]>=22.0",
]

[dependency-groups]
dev = [
    "bandit>=1.8.3",
    "jiwer>=3.0.0",
    "mypy>=1.19.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-cov>=6.1.0",
    "pytest-asyncio>=0.25.0",
    "pytest-mock>=3.14.0",
    "anyio>=4.0.0",
    "ruff>=0.14.13",
]

[project.urls]
Homepage = "https://github.com/w2kr1stn/babel_tower"
Repository = "https://github.com/w2kr1stn/babel_tower"

[project.scripts]
babel = "babel_tower.cli:app"
fmt = "devops:format_code"
test = "devops:test"
clean = "devops:clean"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "N",   # pep8-naming
    "B",   # flake8-bugbear
    "I",   # isort
    "UP",  # pyupgrade
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "B008", # FastAPI Depends() pattern in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"
extraPaths = ["app"]
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = true
reportUnusedVariable = true
reportUnusedFunction = "warning"
reportPrivateUsage = "warning"
reportConstantRedefinition = "error"
reportIncompatibleMethodOverride = "error"
reportIncompatibleVariableOverride = "error"
reportGeneralTypeIssues = "error"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["app"]
addopts = "-ra -q --cov=babel_tower --cov-report=term-missing"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["app/babel_tower"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

[tool.bandit]
exclude_dirs = [".venv", "venv", "tests"]
skips = []
# Fail on Medium and High severity (ISO 27001 compliance)
# Run with: uv run bandit -c pyproject.toml -r .

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchaudio = { index = "pytorch-cpu" }
